requirements_optional.txt
Pillow
PySimpleGUI
PySimpleGUIQt
PySimpleGUIWeb
PyYAML
catpandoc
getostheme
